+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2025-08-31
+
+- Added `shadowSize` parameter to `CardedList` widget to control scroll shadow height
+- Added `shadowColor` parameter to `CardedList` widget to customize scroll shadow color
+- Updated example app to demonstrate shadow customization with a slider control
+
 ## [0.1.0] - 2025-08-31
 
 - Initial public package layout